+++ exposures/manufacturing/manufacturing_sub_exposures/refinement_1/Step3_EDGAR_assign_value.py
@@ -14,7 +14,7 @@ import logging
 LOGGER = logging.getLogger()
 
 from exposures.utils import root_dir
-from utils.s3client import upload_to_s3_bucket
+from nccs.utils.s3client import upload_to_s3_bucket
 
 # Get the root directory
 project_root = root_dir()
