+++ nccs/pipeline/direct/combine_impact_funcs.py
