+++ nccs/pipeline/indirect/indirect.py
