+++ nccs/test/test_analysis.py
@@ -3,10 +3,10 @@ Test the analysis pipeline runs
 """
 
 import unittest
-from analysis import run_pipeline_from_config
-from run_configurations.test.test_config import CONFIG  # change here to test_config if needed
-from utils.folder_naming import get_direct_output_dir
-from utils.delete_results import delete_results_folder
+from nccs.analysis import run_pipeline_from_config
+from nccs.run_configurations.test.test_config import CONFIG  # change here to test_config if needed
+from nccs.utils.folder_naming import get_direct_output_dir
+from nccs.utils.delete_results import delete_results_folder
 
 class TestAnalysisPipeline(unittest.TestCase):
