+++ nccs/utils/folder_naming.py
@@ -1,6 +1,6 @@
 import os
 
-OUTPUT_DIR = os.path.abspath(f"{os.path.dirname(__file__)}/../results")
+OUTPUT_DIR = os.path.abspath(f"{os.path.dirname(__file__)}/../../results")
 
 
 def get_resource_dir():
@@ -8,14 +8,14 @@ def get_resource_dir():
     Returns the absolute path to the exposures directory
     :return:
     """
-    return os.path.abspath(f"{os.path.dirname(os.path.abspath(__file__))}/../resources")
+    return os.path.abspath(f"{os.path.dirname(os.path.abspath(__file__))}/../../resources")
 
 def get_resources_dir():
     """
     Returns the absolute path to the resources directory
     :return:
     """
-    return os.path.abspath(f"{os.path.dirname(os.path.abspath(__file__))}/../resources")
+    return os.path.abspath(f"{os.path.dirname(os.path.abspath(__file__))}/../../resources")
 
 
 def get_output_dir():
