+++ resources/impacts/slr_intermediate/05_summarise_outputs.py
@@ -9,7 +9,7 @@ import itertools
 from functools import reduce, partial
 import traceback
 
-from pipeline.direct.direct import get_sector_exposure
+from nccs.pipeline.direct.direct import get_sector_exposure
 
 aqueduct_metadata_path = Path('.', 'data', 'aqueduct_download_metadata.csv')
 country_bounding_box_path = Path('.', 'data', 'country_bounding_boxes.csv')
