+++ resources/impacts/slr_intermediate/validation_create_haz_exp_mappings.py
@@ -22,7 +22,7 @@ import climada.util.coordinates as u_coord
 import traceback
 import pathos as pa
 
-from pipeline.direct.direct import get_sector_exposure
+from nccs.pipeline.direct.direct import get_sector_exposure
 
 aqueduct_metadata_path = Path('.', 'data', 'aqueduct_download_metadata.csv')
 country_bounding_box_path = Path('.', 'data', 'country_bounding_boxes.csv')
